Few tweeks in place details.

+++ src/screens/PlaceDetail/PlaceDetail.js
@@ -1,16 +1,23 @@
 import React, { Component } from 'react';
-import {View, Image, Text, StyleSheet, TouchableOpacity, Platform, Dimensions} from 'react-native';
+import {View, Image, Text, StyleSheet, Dimensions} from 'react-native';
 import { connect } from 'react-redux';
-import Icon from 'react-native-vector-icons/Ionicons';
 import { Navigation } from 'react-native-navigation';
 import MapView from 'react-native-maps';
 
 import { deletePlace } from '../../store/actions/index'
 
 class PlaceDetail extends Component {
-    placeDeteledHandler = () => {
-        this.props.onDeletePlace(this.props.selectedPlace.key);
-        Navigation.pop(this.props.componentId);
+    constructor(props) {
+        super(props);
+
+        Navigation.events().registerNavigationButtonPressedListener(this.placeDeteledHandler);
+    }
+    
+    placeDeteledHandler = event => {
+        if (event.buttonId === "delete-location-btn" && event.componentId === this.props.componentId) {
+            this.props.onDeletePlace(this.props.selectedPlace.key);
+            Navigation.pop(this.props.componentId);
+        }        
     }
     
     render() {
@@ -24,18 +31,11 @@ class PlaceDetail extends Component {
         return (
             <View style={styles.container}>
                 <View>
+                    <Text style={styles.placeName}>{this.props.selectedPlace.name}</Text>
                     <Image style={styles.placeImage} source={this.props.selectedPlace.image} />
                     <MapView style={styles.map} initialRegion={location}>
                         <MapView.Marker coordinate={location} />
                     </MapView>
-                    <Text style={styles.placeName}>{this.props.selectedPlace.name}</Text>
-                </View>
-                <View>
-                    <View style={styles.deleteButton} >
-                        <TouchableOpacity onPress={this.placeDeteledHandler}>
-                            <Icon size={30} name={Platform.OS === 'android' ? "md-trash" : "ios-trash"} color="red" />
-                        </TouchableOpacity>
-                    </View>
                 </View>
             </View>
         );
@@ -49,6 +49,7 @@ const styles = StyleSheet.create({
     placeImage: {
         width: "100%",
         height: 200,
+        marginBottom: 10,
     },
     map: {
         width: "100%",
